gateway/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
distlib==0.3.7
ecdsa==0.18.0
fastapi==0.103.1
fastapi-oidc==0.0.9
filelock==3.12.4
grpcio==1.58.0
grpcio-tools==1.58.0
h11==0.14.0
idna==3.4
jwt==1.3.1
passlib==1.7.4
platformdirs==3.10.0
protobuf==4.24.3
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.24.5
